requirements.txt
annotated-types==0.6.0
anyio==4.2.0
autopep8==2.0.4
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.109.2
h11==0.14.0
idna==3.6
mysql-connector-python==8.3.0
pycodestyle==2.11.1
pydantic==2.6.1
pydantic_core==2.16.2
sniffio==1.3.0
starlette==0.36.3
tomli==2.0.1
typing_extensions==4.9.0
uvicorn==0.27.0.post1
